Dedupe requests that are not explicitly no-cache
If a location is known to be _not_ no-cache, or its cache status is
unknown (so we're holding back subsequent requests) and the first
request comes back without no-cache, we'll apply the response to all
the requests that came in the meantime.

If we're expecting a response to not be no-cache, but it turns up
actually not cacheable after all, then we'll issue any requests that
may have been delayed pending the assumed-cacheable response.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>