Add cache header parsing, with expiry
I wrote this with the intent of distinguishing between uncacheable
resources and resources that we could dedupe.  But it turns out to be
just about as easy to write a full expiry implementation.

There's more that needs doing -- I don't test anything with a max-age
and we don't actually expire entries out of the cache at the moment.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>