Run Jest against JS, not TS
It turns out to be much easier to convince tsc to emit a coherent build
of modules that can reference each other than it is Jest to discover
what's going on.

We still have two modules, src and test, but now we refer to them
through "imports" defined in package.json.  This lets us relocate test
wherever we want to, and still have Jest consume the result.  Jest now
has no knowledge of TS whatsoever, and operates on plain ESM modules.
And we can use Node16 modules and resolution without breaking anything.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>